refactor(@damons-ui/css-core): rename set into pseudo
