+++ packages/core/css/src/index.ts
@@ -1,7 +1,7 @@
 export * from "./unit";
 export * from "./types";
 export * from "./utils";
-export * from "./set";
+export * from "./pseudo";
 export * from "./text";
 export * from "./box";
 export * from "./position";
