+++ packages/core/css/src/pseudo.ts
